FIX: Fix style
