+++ scripts/check_version.py
@@ -12,6 +12,7 @@ with open('templates/config.toml', 'r') as fid:
 # Check config secret
 if 'auth' in mapping and 'shared_secret' in mapping['auth']:
     if mapping['auth']['shared_secret'] == template['auth']['shared_secret']:
+
         sys.stderr.write('\nWarning: Your shared secret has not been changed from its default.\n'
                          'Your system will be insecure until you change auth.shared_secret in config.toml.\n')
 
@@ -19,7 +20,6 @@ if 'auth' in mapping and 'shared_secret' in mapping['auth']:
 # Check config version
 if ('flywheel' in mapping and 'version' in mapping['flywheel'] and
         isinstance(mapping['flywheel']['version'], (int, float))):
-
     version = mapping['flywheel']['version']
     expected = template['flywheel']['version']
 
@@ -37,7 +37,7 @@ if ('flywheel' in mapping and 'version' in mapping['flywheel'] and
 else:
     sys.stderr.write('\nNo version key found in your config.toml.\n'
                      'This likely means your configuration is very old, or invalid.\n'
-                     'Delete your config.toml to generate a new default configuration.')
+                     'Delete your config.toml to generate a new default configuration.\n')
     sys.exit(2)
