+++ scripts/template.py
@@ -4,10 +4,9 @@ import os
 import sys
 import pystache
 import toml
-import re
 
 mapP = sys.argv[1]
-temP  = sys.argv[2]
+temP = sys.argv[2]
 repairPythonTypes = True if len(sys.argv) > 3 else False
 
 # Read user and static settings
@@ -19,19 +18,18 @@ mapping = static.copy()
 mapping.update(settings)
 
 # Hackaround for any template that needs an absolute path
-mapping["absPath"]  = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+mapping["absPath"] = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 result = pystache.render(open(temP, 'r').read(), mapping)
 
 # Horrible hackaround for python booleans being silly
 # Better way would be to control pystache template rendering
 if repairPythonTypes:
-	result = result.replace("True", "true").replace("False", "false")
+    result = result.replace("True", "true").replace("False", "false")
 
 print result
 
 try:
-	sys.stdout.flush()
-except:
-	pass
-
+    sys.stdout.flush()
+except Exception:
+    pass
